+++ src/dico/command/MenuItemSearchStart.java
@@ -4,7 +4,7 @@ import dico.Menu;
 
 public class MenuItemSearchStart extends MenuItemSearch {
 	
-	public MenuItemSearchStart(Menu menu, short id) {
+	public MenuItemSearchStart(Menu menu, Integer id) {
 		
 		super(menu, "Rechercher tous les mots qui commencent par une valeur", id);
 	}
